+++ srcs/utils.c
@@ -12,9 +12,9 @@
 
 #include "../includes/minishell.h"
 
-int	max_len(char *str, int nb)
+int	max_len(char *str, size_t nb)
 {
-	if (ft_strlen(str) > (size_t)nb)
+	if (ft_strlen(str) > nb)
 		return (ft_strlen(str));
 	return (nb);
 }
